requirements.txt
docker==7.1.0
PyYAML==6.0.3
watchdog==6.0.0
